+++ src/main/java/lotto/domain/LottoResult.java
@@ -1,12 +1,10 @@
 package lotto.domain;
 
-import java.text.DecimalFormat;
-import java.util.HashMap;
+import lotto.domain.vo.LottoMoney;
+
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 public class LottoResult {
     private final LottoMoney lottoMoney;
