+++ src/main/java/lotto/domain/number/LottoNumberGenerator.java
@@ -1,6 +1,5 @@
-package lotto.domain;
+package lotto.domain.number;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
