+++ src/test/java/lotto/domain/number/LottoNumberTest.java
@@ -1,12 +1,11 @@
-package lotto.domain;
+package lotto.domain.number;
 
+import lotto.domain.number.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static org.junit.jupiter.api.Assertions.*;
 
 class LottoNumberTest {
     @DisplayName("Lotto 번호가 1 ~ 45 사이의 값이 아니면 IllegalArgumentException Throw")
