+++ src/test/java/lotto/domain/vo/LottoMoneyTest.java
@@ -1,14 +1,13 @@
-package lotto.domain;
+package lotto.domain.vo;
 
+import lotto.domain.vo.LottoMoney;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static org.junit.jupiter.api.Assertions.*;
 
 class LottoMoneyTest {
